+++ spec/spec_helper.rb
@@ -4,6 +4,8 @@ require "ma"
 require 'pry' unless ENV['CI']
 
 RSpec.configure do |config|
+  config.filter_run_when_matching :focus
+  config.run_all_when_everything_filtered = true
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = ".rspec_status"