+++ spec/unit/subscriber_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe Ma::Subscriber do
       stub_const('UserPromotedEvent', double(name: 'UserPromotedEvent'))
 
       subscriber = Class.new do
-        include Ma::Subscriber
+        include Ma::Subscriber.new
 
         on(UserPromotedEvent) {}
       end.new
@@ -17,12 +17,34 @@ RSpec.describe Ma::Subscriber do
       end
 
       subscriber = Class.new do
-        include Ma::Subscriber
+        include Ma::Subscriber.new
 
         on(UserPromotedEvent) {}
       end.new
 
       expect(subscriber.method(:UserPromotedEvent).arity).to eq(1)
+    end
+
+    describe 'when :async option passed' do
+      describe 'and async broadcaster is defined' do
+        it 'uses async broadcaster' do
+          async_broadcaster = double
+          async_broadcaster_class = double(new: async_broadcaster)
+
+          stub_const('UserPromotedEvent', double('UserPromotedEvent', new: double('user_promoted_event'), name: 'UserPromotedEvent'))
+          stub_const('WisperNext::Subscriber::AsyncBroadcaster', async_broadcaster)
+
+          expect(async_broadcaster).to receive(:new)
+
+          subscriber = Class.new do
+            include Ma::Subscriber.new(:async)
+
+            on(UserPromotedEvent) {}
+          end.new
+
+          subscriber.send(:UserPromotedEvent, UserPromotedEvent.new)
+        end
+      end
     end
   end
 end